+++ test/libsolidity/syntaxTests/globalFunctions/sha3_override.sol
@@ -1,0 +1,11 @@
+contract C
+{
+	function sha3() public pure returns (bool) {
+		return true;
+	}
+	function f() public pure returns (bool) {
+		return sha3();
+	}
+}
+// ----
+// Warning: (14-76): This declaration shadows a builtin symbol.